Improve MPRIS2 implementation
Now it works fine with the gnome media player indicator extension:
  - album art is now properly displayed
  - fix postion slider information, Gnome require the seeked signal to be emitted. There is one annoyance left, gnome automatically update the timer using it's own timer and there are sometimes where the two timer are not synced correctly. This is a gnome issue.

+++ lib/include/mellowplayer/player.h
@@ -120,7 +120,7 @@ signals:
      *
      * \param position New song position [µs]
      */
-    void positionChanged(int position);
+    void positionChanged(qlonglong position);
 
 private slots:
     void onArtReady(const QString& filePath);
